// e2e/enums/component-enum/text-field.enum.ts
export enum TextField {
  EMAIL_ADDRESS = "email-address",
  PASSWORD = "password",
  FULLNAME = "fullname",
  WORKEMAIL = "work-email",
  COMPANY = "company",
  CONFIRMPASSWORD = "confirm-password",
  PRODUCT_NAME = "product-name",
  PRODUCT_CODE = "product-code",
  CATEGORY = "category",
  PRICE = "price",
  STOCK = "stock",
}